adding in a random chance to delete the message.

+++ events/dabbleDelete.js
@@ -6,12 +6,16 @@ module.exports = {
     if (!(message.channelId in global.messageHistory))
       global.messageHistory[message.channelId] = message.author.id;
     else {
+      const randValue = 100 * Math.random();
       if (
         global.messageHistory[message.channelId] === message.author.id &&
         message.author.id === global.dabUserId &&
+        randValue > 100 - global.dabbleRatio &&
         global.isDabbleEnabled
       ) {
-        console.log(`Deleting message: ${message.content}`);
+        console.log(
+          `Deleted Message: ${message.content} sent to channel: (${message.channelId})`
+        );
         message.delete();
       } else global.messageHistory[message.channelId] = message.author.id;
     }
